smal styling and structure improvements
