+++ static_data/theme.tsx
@@ -236,6 +236,13 @@ const theme = extendTheme({
           } as CSSObject
         }
       }
+    },
+    Link: {
+      baseStyle: {
+        _hover: {
+          textDecoration: 'none'
+        }
+      }
     }
   }
 });
